+++ src/remakeDir.ts
@@ -1,12 +1,18 @@
 import * as fs from 'fs-extra';
+import fileAccess from './fileAccess';
 
 async function remakeDir(dir: string): Promise<void> {
   try {
-    await fs.access(dir);
-    await fs.remove(dir);
-    await fs.mkdirp(dir);
+    const access = await fileAccess(dir);
+
+    if (access) {
+      await fs.remove(dir);
+      await fs.mkdirp(dir);
+    } else {
+      await fs.mkdirp(dir);
+    }
   } catch (err) {
-    await fs.mkdirp(dir);
+    console.error('error while remaking dir');
   }
 }
